+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0
 
 - Improved TypeScript version independence
+- Fixed importing types of hidden submodules (#9)
 - Added compatibility with recent (or any) TypeScript release
 - Added support for optional entries in tuples